python/requirements.txt
# Core dependencies
numpy>=1.20.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=3.20.0

# TensorFlow - the tensorflow-macos package is optimized for Apple Silicon
tensorflow-macos>=2.10.0; sys_platform == 'darwin' and platform_machine == 'arm64'
tensorflow>=2.10.0; (sys_platform != 'darwin' or platform_machine != 'arm64')

# Metal plugin for Apple Silicon GPU acceleration
tensorflow-metal>=0.8.0; sys_platform == 'darwin' and platform_machine == 'arm64' 